fix: limit delete secret prompt to when there is a secret

+++ internal/pkg/cli/pipeline_delete.go
@@ -118,8 +118,10 @@ func (o *deletePipelineOpts) Ask() error {
 
 // Execute deletes the secret and pipeline stack.
 func (o *deletePipelineOpts) Execute() error {
-	if err := o.deleteSecret(); err != nil {
-		return err
+	if o.PipelineSecret != "" {
+		if err := o.deleteSecret(); err != nil {
+			return err
+		}
 	}
 
 	if err := o.deleteStack(); err != nil {
